requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
duckdb>=0.9.0
plotly>=5.0.0
python-dotenv>=1.0.0
openai>=1.0.0
shiny>=0.8.0
